src/requirements.txt
google-cloud-secret-manager==2.11.1
google-cloud-storage==2.3.0